Noobs can now be instantiated along with their very own health bar.
Whenever a Noob is instantiated, a health bar is also instantiated.

I created a new prefab for the Noobs where they have a health bar gameObject attached to them. Please look at the Resources folder to take a look at it.

I also made a prefab for the health bar gameObject by itself.